src/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements-dev.in
#
attrs==21.2.0
    # via pytest
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via requests
deprecation==2.1.0
    # via testcontainers
docker==5.0.3
    # via testcontainers
factory-boy==3.2.1
    # via -r requirements-dev.in
faker==9.8.3
    # via factory-boy
greenlet==1.1.2
    # via sqlalchemy
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==21.2
    # via
    #   deprecation
    #   pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pymysql==1.0.2
    # via testcontainers
pyparsing==2.4.7
    # via packaging
pytest==6.2.3
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via faker
requests==2.25.1
    # via docker
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.26
    # via testcontainers
testcontainers[mysql]==3.4.2
    # via -r requirements-dev.in
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
urllib3==1.26.7
    # via requests
websocket-client==1.2.1
    # via docker
wrapt==1.13.3
    # via testcontainers
